requirements.txt
apache-beam[gcp]==2.53.0
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.11.0
